GenBI/requirements.txt
# FastAPI và server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# LLM
openai==1.3.5
google-generativeai==0.3.1

# Database và xử lý dữ liệu
duckdb==0.9.2
pandas==2.1.3
sqlparse==0.4.4

# Visualization
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.0

# Frontend
streamlit==1.28.2

# Utils
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0